Transformers refactor started.
